+++ src/com/github/mjkuranda/retrographicsmanager/commandsystem/commands/GeneratePaletteCommand.java
@@ -13,21 +13,23 @@ import java.time.LocalDateTime;
 
 public class GeneratePaletteCommand implements Command {
 
-    private String[] lineArgs;
+    private String paletteName;
+
+    private PaletteTypes type;
+
+    private boolean isBrighter;
+
+    private int scale;
 
     public GeneratePaletteCommand(String[] lineArgs) {
-        this.lineArgs = lineArgs;
+        this.type = "block".equals(lineArgs[2]) ? PaletteTypes.BLOCK : PaletteTypes.RECTANGLE;
+        this.isBrighter = "light".equals(lineArgs[1]);
+        this.scale = lineArgs.length > 3 && lineArgs[3].matches("^-?\\d+$") ? Integer.parseInt(lineArgs[3]) : 1;
+        this.paletteName = getPaletteName(lineArgs);
     }
 
     @Override
     public void execute() {
-        String dateTime = LocalDateTime.now().toString().substring(0, 19).replaceAll(":", "-");
-        String paletteName = dateTime + "-palette-" + lineArgs[1] + "-" + lineArgs[2];
-
-        if (lineArgs.length > 3) {
-            paletteName = lineArgs[3];
-        }
-
         if (!new File("dat/output").exists()) {
             try {
                 Files.createDirectories(Paths.get("dat/output"));
@@ -36,9 +38,13 @@ public class GeneratePaletteCommand implements Command {
             }
         }
 
-        PaletteTypes type = "block".equals(lineArgs[2]) ? PaletteTypes.BLOCK : PaletteTypes.RECTANGLE;
-        boolean isBrighter = "light".equals(lineArgs[1]);
         Palette palette = PaletteFactory.get(type, isBrighter);
-        PaletteWriter.save(palette, paletteName, 1);
+        PaletteWriter.save(palette, paletteName, scale);
+    }
+
+    private String getPaletteName(String[] lineArgs) {
+        String dateTime = LocalDateTime.now().toString().substring(0, 19).replaceAll(":", "-");
+
+        return dateTime + "-palette-" + lineArgs[1] + "-" + lineArgs[2] + "-" + this.scale;
     }
 }