+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.0] - 2023-05-20
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): Option `dirType` to `ListCommand`.
+- [@mjkuranda](https://github.com/mjkuranda): Scale parameter to `GeneratePaletteCommand`.
+
+### Changed
+- [@mjkuranda](https://github.com/mjkuranda): Optimized `ListCommand` and `GeneratePaletteCommand` commands.
+
 ## [3.3.0] - 2023-05-20
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): `generate`/`gen` command to generate light or bright palettes with a specific name but scale 1.